refactored the signOut and moved signIn and signOut into auth package
